Deepen favorite articles module
